Refactoring aller Funktionen
